requirements-core.txt
# Core packages only (minimal installation)
# For users behind firewall or with slow network

# Essential libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0

# Machine learning models
lightgbm>=4.0.0
xgboost>=2.0.0

# Imbalanced learning
imbalanced-learn>=0.11.0

# Ensemble selection (optional, can install later)
# deslib>=0.3.7
